gateway-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
azure-servicebus==7.11.4
pydantic==2.5.3
